@@ -1,30 +1,28 @@
 import 'package:flutter/material.dart';
+import 'package:rallis/utils/constant/assets_const.dart';
 import 'package:rallis/utils/constant/color_const.dart';
-import 'package:rallis/utils/theme/text_style.dart';
 
 ThemeData lightThemeData(BuildContext context) {
   return ThemeData(
     primaryColor: ColorConst.appColor,
-    colorScheme: ColorScheme.light(
-        primary: ColorConst.appColor, secondary: ColorConst.appColor),
     indicatorColor: Colors.white,
     splashColor: Colors.white24,
     splashFactory: InkRipple.splashFactory,
-    fontFamily: "poppins",
+    fontFamily: AssetsConst.poppinsFonts,
     shadowColor: Theme.of(context).disabledColor,
     dividerColor: const Color.fromARGB(255, 112, 112, 112),
     canvasColor: Colors.white,
-    backgroundColor: const Color.fromARGB(255, 255, 255, 255),
     scaffoldBackgroundColor: const Color.fromARGB(255, 245, 245, 245),
-    errorColor: ColorConst.redColor,
     visualDensity: VisualDensity.adaptivePlatformDensity,
     tabBarTheme: const TabBarTheme(
-      unselectedLabelStyle: AppTextStyle.normalGrey12,
       labelColor: Colors.black,
       unselectedLabelColor: Colors.grey,
       indicatorSize: TabBarIndicatorSize.tab,
-      labelStyle: AppTextStyle.normalBlack12,
     ),
+    colorScheme: const ColorScheme.light(
+            primary: ColorConst.appColor, secondary: ColorConst.appColor)
+        .copyWith(surface: const Color.fromARGB(255, 255, 255, 255))
+        .copyWith(error: ColorConst.redColor),
   );
 }
 
@@ -32,16 +30,16 @@ ThemeData darkThemeData(BuildContext context) {
   return ThemeData(
     scaffoldBackgroundColor: Colors.grey.shade900,
     primaryColor: ColorConst.appColor,
-    colorScheme: const ColorScheme.dark(),
     indicatorColor: Colors.white,
     splashColor: Colors.white24,
     splashFactory: InkRipple.splashFactory,
-    fontFamily: "poppins",
+    fontFamily: AssetsConst.poppinsFonts,
     shadowColor: Theme.of(context).disabledColor,
     dividerColor: const Color(0xff707070),
     canvasColor: Colors.white,
-    backgroundColor: Colors.black,
-    errorColor: ColorConst.redColor,
     visualDensity: VisualDensity.adaptivePlatformDensity,
+    colorScheme: const ColorScheme.dark()
+        .copyWith(surface: Colors.black)
+        .copyWith(error: ColorConst.redColor),
   );
 }
